pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gibran"
version = "0.0.2"
description = "Gibran: governed metric layer + non-LLM NL-to-SQL over DuckDB. 19 metric primitives (incl. cohort_retention, funnel, multi_stage_filter); identity-aware row + column governance; audit-log redaction; pattern-template NL with structural no-hallucination guarantee."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Adam Ezzat" }]
keywords = [
    "analytics",
    "metric-layer",
    "semantic-layer",
    "governance",
    "duckdb",
    "sql",
    "dsl",
    "audit",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "duckdb>=1.0",
    "pyyaml>=6.0",
    "typer>=0.12",
    "pyjwt[crypto]>=2.8",
    "pydantic>=2.0",
    "sqlglot>=23",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
# Phase 5A.2: optional PostgresEngine. Install via `pip install gibran[postgres]`.
# Kept out of the default install so headless / DuckDB-only deployments
# don't pull psycopg's binary wheel.
postgres = ["psycopg[binary]>=3.1"]
# Phase 5A.3: optional SnowflakeEngine. Install via `pip install gibran[snowflake]`.
# Snowflake adapter code is in-tree; running against a real Snowflake account
# requires a paid Snowflake subscription. See tests/test_snowflake_engine_integration.py
# for the always-skip-without-credentials integration test stub.
snowflake = ["snowflake-connector-python>=3.0"]
# Phase 5A.4: optional BigQueryEngine. Install via `pip install gibran[bigquery]`.
# BigQuery adapter code is in-tree; running against a real BigQuery project
# requires a GCP project with billing enabled. See tests/test_bigquery_engine_integration.py
# for the always-skip-without-credentials integration test stub.
bigquery = ["google-cloud-bigquery>=3.0"]
# Phase 4A: optional UI HTTP server. Install via `pip install gibran[ui]`.
# Headless / CLI-only deployments don't pull FastAPI / uvicorn.
ui = ["fastapi>=0.110", "uvicorn>=0.27"]

[project.urls]
Homepage = "https://github.com/AdamEzzat1/gibran"
Repository = "https://github.com/AdamEzzat1/gibran"
Issues = "https://github.com/AdamEzzat1/gibran/issues"

[project.scripts]
gibran = "gibran.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/gibran"]
# Phase 4B: pack the React build output into the wheel so installs via
# `pip install gibran[ui]` get the SPA without needing a Node toolchain.
# The frontend/ directory is a separate npm project; `npm run build`
# emits to src/gibran/ui/static/, which Hatch's package globbing picks
# up automatically (no force-include needed since it lives under the
# gibran package). If the build hasn't run, the directory is empty
# and the server falls back to a JSON placeholder.
artifacts = ["src/gibran/ui/static/**"]
